+++ test/test-help.sh
@@ -8,6 +8,9 @@ cmpOsLog helpMain
 $CMD sample     > ${TMP}/helpSample.txt	 2>&1
 cmpOsLog helpSample
 
+$CMD degrade    > ${TMP}/helpDegrade.txt	 2>&1
+cmpOsLog helpDegrade
+
 $CMD compare   > ${TMP}/helpCompare.txt 2>&1
 cmpOsLog helpCompare
 
@@ -16,6 +19,9 @@ cmpOsLog helpReindex
 
 $CMD quantize   > ${TMP}/helpQuantize.txt 2>&1
 cmpOsLog helpQuantize
+
+$CMD sequence   > ${TMP}/helpSequence.txt 2>&1
+cmpOsLog helpSequence
 
 $CMD badcmd     > ${TMP}/helpBadCmd.txt 2>&1
 cmpOsLog helpBadCmd