+++ test/test-sample-grid.sh
@@ -32,7 +32,7 @@ cmp ${TMP}/${OUT}.ply ${REFS}/${OUT}.ply
 # test degenerate triangle
 OUT=sample_grid_degenerate
 $CMD sample -i ${DATA}/degenerate.obj -o ${TMP}/${OUT}.ply --mode grid --hideProgress --gridSize 10 > ${TMP}/${OUT}.txt 2>&1
-fileHasString ${TMP}/${OUT}.txt "Skipped 1 triangles" 1
+fileHasString ${TMP}/${OUT}.txt "Skipped 1 degenerate triangles" 1
 
 # external dataset
 if [ "$1" == "ext" ];