+++ test/test-sample-sdiv.sh
@@ -32,7 +32,7 @@ $CMD sample -i ${DATA}/sphere.obj -o ${TMP}/${OUT}.ply --mode sdiv --hideProgres
 # test degenerate triangle
 OUT=sample_sdiv_degenerate
 $CMD sample -i ${DATA}/degenerate.obj -o ${TMP}/${OUT}.ply --mode sdiv --hideProgress > ${TMP}/${OUT}.txt 2>&1
-fileHasString ${TMP}/${OUT}.txt "Skipped 1 triangles" 1
+fileHasString ${TMP}/${OUT}.txt "Skipped 1 degenerate triangles" 1
 
 # external dataset
 if [ "$1" == "ext" ];